requirements.txt
# Core Framework
Flask>=3.0.0
Werkzeug>=3.0.0

# Database
Flask-SQLAlchemy>=3.0.0
Flask-Migrate>=4.0.0
psycopg2-binary>=2.9.0

# Authentication
Flask-JWT-Extended>=4.5.0

# Configuration
python-dotenv>=1.0.0

# Optional: Redis (uncomment if needed)
# redis>=5.0.0


# HTTP Client (for Backend App)
requests>=2.31.0
